+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Versions
 
-## 2.18.2 (??)
+## 2.18.2 (18 Aug 2022)
 - Make the call to Check Wikipedia non-blocking
 - Include in the edit summary all the applied replacements
 - Count replacements by type without grouping by page